matz/sngan/requirements.txt
chainer==3.3.0
tensorflow-gpu==1.2.0
numpy==1.11.1
cython==0.27.2
cupy==2.0.0
scipy==0.19.0
pillow==4.3.0
pyyaml==3.12
h5py==2.7.1
